requirements.txt
Flask==2.1.2
Pillow==9.2.0
PyPDF2==2.4.2
pytest==7.1.2
fpdf==1.7.2
boto3==1.24.50
botocore==1.27.50
